# server/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart100X API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://smart100x-w9a4.vercel.app",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:5001",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/chat": {
      "post": {
        "description": "This endpoint allows you to create data for the Chat resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Facebook",
          "Instagram",
          "Twitter"
        ],
        "summary": "Create Chat"
      }
    },
...
    "/api/support/chat": {
      "post": {
        "description": "This endpoint allows you to create data for the Support Chat resource. Authentication may be required depending on the route.",
        "responses": {
          "0": {
            "description": ""
          },
          "1": {
            "description": ""
          },
          "4": {
            "description": ""
          },
          "5": {
            "description": ""
          },
          "7": {
            "description": ""
          },
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "example": "any"
                  },
                  "history": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Facebook",
          "Instagram",
          "Twitter"
        ],
        "summary": "Create Support Chat"
      }
    },
    "/api/support/contact": {
      "post": {
        "description": "This endpoint allows you to create data for the Support Contact resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "subject": {
                    "example": "any"
                  },
                  "message": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Create Support Contact"
      }
    },
...
    "/chat": {
      "post": {
        "description": "This endpoint allows you to create data for the  Chat resource. Authentication may be required depending on the route.",
        "responses": {
          "0": {
            "description": ""
          },
          "1": {
            "description": ""
          },
          "4": {
            "description": ""
          },
          "5": {
            "description": ""
          },
          "7": {
            "description": ""
          },
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "example": "any"
                  },
                  "history": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Facebook",
          "Instagram",
          "Twitter"
        ],
        "summary": "Create  Chat"
      }
    },
    "/contact": {
      "post": {
        "description": "This endpoint allows you to create data for the  Contact resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "subject": {
                    "example": "any"
                  },
                  "message": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Create  Contact"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}